Feat(#42): @Authenticated 적용을 위한 Resolver 작성
