--- application/wypl-core/src/main/java/com/wypl/wyplcore/auth/domain/AuthMember.java
@@ -1,14 +1,0 @@
-package com.wypl.wyplcore.auth.domain;
-
-import com.wypl.jpacalendardomain.calendar.data.ConvertibleScheduleInfo;
-
-public record AuthMember(
-	long id
-) implements ConvertibleScheduleInfo {
-
-	@Override
-	public Long getCreatorId() {
-		return id();
-	}
-
-}
