+++ client/google-oauth-client/src/main/java/com/wypl/googleoauthclient/annotation/Authenticated.java
@@ -1,4 +1,4 @@
-package com.wypl.wyplcore.auth.annotation;
+package com.wypl.googleoauthclient.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
