+++ client/google-oauth-client/src/main/java/com/wypl/googleoauthclient/domain/AuthMember.java
@@ -1,0 +1,10 @@
+package com.wypl.googleoauthclient.domain;
+
+public record AuthMember(
+	long id,
+	String accessToken
+) {
+	public static AuthMember of(long id, String accessToken) {
+		return new AuthMember(id, accessToken);
+	}
+}
